Require password after some minutes
The wallet's private keys are currently exposed for the entire duration of the
account being unlocked. While this is a nice user experience it is not security
oriented. As an alternative, the wallet's private keys will be available for
additional signing for a certain amount of adjustable time after a transaction.
Each subsequent transaction will reset the timer. This enables users to continue
to view their funds, while keeping their private keys safe.